Extract HttpPusher to its own file
